validator/pyproject.toml
[build-system]
requires = ["setuptools >= 75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-maxxing-validator"
description = "The validator which checks models and checkpoints provided by miners"
requires-python = ">=3.10,<3.13"
version = "5.5.0"
dependencies = [
    "edge-maxxing-base==1.0.0",
    "opentelemetry-api>=1.28.2",
    "opentelemetry-sdk>=1.28.2",
    "opentelemetry-exporter-otlp>=1.28.2",
    "opentelemetry-distro>=0.49b2",
    "wandb>=0.18.7",
    "opentelemetry-instrumentation-asyncio>=0.49b2",
    "opentelemetry-instrumentation-dbapi>=0.49b2",
    "opentelemetry-instrumentation-logging>=0.49b2",
    "opentelemetry-instrumentation-sqlite3>=0.49b2",
    "opentelemetry-instrumentation-threading>=0.49b2",
    "opentelemetry-instrumentation-urllib>=0.49b2",
    "opentelemetry-instrumentation-wsgi>=0.49b2",
    "opentelemetry-instrumentation-aiohttp-client>=0.49b2",
    "opentelemetry-instrumentation-aiohttp-server>=0.49b2",
    "opentelemetry-instrumentation-fastapi>=0.49b2",
    "opentelemetry-instrumentation-grpc>=0.49b2",
    "opentelemetry-instrumentation-httpx>=0.49b2",
    "opentelemetry-instrumentation-jinja2>=0.49b2",
    "opentelemetry-instrumentation-requests>=0.49b2",
    "opentelemetry-instrumentation-starlette>=0.49b2",
    "opentelemetry-instrumentation-system-metrics>=0.49b2",
    "opentelemetry-instrumentation-tortoiseorm>=0.49b2",
    "opentelemetry-instrumentation-urllib3>=0.49b2",
    "fastapi>=0.115.5",
]

[tool.uv.sources]
edge-maxxing-base = { path = "../base", editable = true }

[dependency-groups]
dev = [
    "pytype==2024.10.11"
]

[tool.pytype]
inputs = [
    "base_validator",
    "weight_setting",
    "submission_tester",
]

[project.scripts]
start_validator = "weight_setting.validator:main"

[tool.setuptools]
packages = [
    "base_validator",
    "weight_setting",
    "submission_tester",
]
